app/schemas/link_list_response_schema.py
from app.schemas.link_detail_schema import LinkDetailSchema
from pydantic import BaseModel


class LinksResponseSchema(BaseModel):
    count: int
    limit: int
    offset: int
    items: list[LinkDetailSchema]
