pyproject.toml
[tool.poetry]
name = "url-link-detector"
version = "0.1.0"
description = ""
authors = ["aleksandr <alsv@smena.space>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.2"
uvicorn = "^0.31.1"
dependency-injector = "^4.42.0"
pydantic-settings = "^2.5.2"
sqlalchemy = "^2.0.35"
alembic = "^1.13.3"
python-dotenv = "^1.0.1"
asyncpg = "^0.29.0"
psycopg2-binary = "^2.9.9"
orjson = "^3.10.7"
httpx = "^0.27.2"
celery = "^5.4.0"
redis = "^5.1.1"
flower = "^2.0.1"
ipython = "^8.28.0"
sqladmin = {extras = ["full"], version = "^0.19.0"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.poetry.group.lint.dependencies]
isort = "^5.13.2"
flake8 = "^7.0.0"
ruff = "^0.4.3"
pre-commit = "^3.7.0"

[tool.flake8]
extend-ignore = """
    A003,
"""

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
line_length = 120
lines_after_imports = 2
lines_between_sections = 1
lines_between_types = 0
force_alphabetical_sort_within_sections = true
ensure_newline_before_comments = true
combine_as_imports = true
combine_star = true
skip_gitignore = true
balanced_wrapping = true
force_grid_wrap = 2
src_paths = ["core", ]
skip_glob = [
    ".git",
    ".idea",
    ".vscode",
    "venv",
    ".venv",
]
known_fastapi = ["fastapi", "starlette"]
known_first_party = ["application", "domain", "infra", "logic", "settings"]
sections = [
    "FUTURE",
    "STDLIB",
    "FASTAPI",
    "THIRDPARTY",
    "FIRSTPARTY",
    "LOCALFOLDER",
]